feat: added support for `//MAIN`
The main class for a script can now be explicitly set by using the new
`//MAIN` tag. This overrides the default detection that Jbang  performs
and is assumed to be correct, no validation is done.
